Merge/#61 - user테이블 수정과 pk값 변경 관련 merge

+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/chat/dto/ChatRequestDto.java
@@ -10,7 +10,7 @@ public class ChatRequestDto {
 
 
 
-    private Integer senderId;
+    private Long senderId;
 
     private String message;
 }
